add: annotation db util
