+++ src/redux/sagas/utils/index.ts
@@ -10,6 +10,7 @@ export * from './enableExtensionHelpers';
 export * from './refreshExtension';
 export * from './metadataChangelog';
 export * from './ipfs';
+export * from './annotations';
 
 // export * from './updateColonyDisplayCache';
 // export { updateMotionValues } from './updateMotionValues';
